Add files via upload
 sample of Uber Bus trips for the month of February
2021. In the Uber Bus experience, the rider has the option to either request the trip in advance
or on the fly. After requesting the trip, the rider sees on the map the pickup point (bus station)
closest to him/her, and the dropoff point (bus station) closest to his/her destination and the
closest time for pickup. After confirming the trip, the trip gets logged with all the corresponding
information in the dataset file attached. The trip is mostly completed unless the rider decides to
cancel the trip or the driver wasn’t able to make it and cancel the trip from his / her end. Finally,
a rider can purchase a pass / subscription on the bus, in such cases the trip will be marked as a
subscription trip.

1) Created a mysql database for the data that contains two tables, one for the
trip data and one for the vehicle information (vehicle id and capacity), and have populated the
tables with the data from the CSV file.
2) Created a python notebook to create reports for the business metrics with the
appropriate SQL queries needed to extract the data for each report.